# main.go
package main

import (
	"log"
	"net/http"
	"os"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/joho/godotenv"

	"virtual-cuppa-be/config"
	"virtual-cuppa-be/handlers"
	"virtual-cuppa-be/middleware"
	"virtual-cuppa-be/repositories"
	"virtual-cuppa-be/scheduler"
	"virtual-cuppa-be/services"
)

func main() {
	if err := godotenv.Load(); err != nil {
		log.Println("No .env file found, using default values")
	}

	config.ConnectDatabase()

	userRepo := repositories.NewUserRepository(config.DB)
	orgRepo := repositories.NewOrganisationRepository(config.DB)
	tagRepo := repositories.NewTagRepository(config.DB)
	matchRepo := repositories.NewMatchRepository(config.DB)
	matchHistoryRepo := repositories.NewMatchHistoryRepository(config.DB)
	matchFeedbackRepo := repositories.NewMatchFeedbackRepository(config.DB)
	userAvailConfigRepo := repositories.NewUserAvailabilityConfigRepository(config.DB)
	emailService := services.NewEmailService()
	matchService := services.NewMatchService(matchRepo, matchHistoryRepo, matchFeedbackRepo, userRepo, userAvailConfigRepo, emailService)
	authService := services.NewAuthService(userRepo, emailService, matchService)
	userService := services.NewUserService(userRepo, orgRepo, tagRepo, emailService)
	orgService := services.NewOrganisationService(orgRepo)
	userAvailConfigService := services.NewUserAvailabilityConfigService(userAvailConfigRepo, userRepo)
	authHandler := handlers.NewAuthHandler(authService)
	userHandler := handlers.NewUserHandler(userService)
	orgHandler := handlers.NewOrganisationHandler(orgService, userService)
	userAvailConfigHandler := handlers.NewUserAvailabilityConfigHandler(userAvailConfigService, matchService)

	// Start match scheduler
	matchScheduler := scheduler.NewMatchScheduler(matchService, orgRepo)
	matchScheduler.Start()

	matchHandler := handlers.NewMatchHandler(matchService, matchScheduler)
	feedbackHandler := handlers.NewMatchFeedbackHandler(matchService)

	router := gin.Default()

	allowedOriginsEnv := os.Getenv("CORS_ALLOWED_ORIGINS")
	if allowedOriginsEnv == "" {
		allowedOriginsEnv = "https://virtual-cappa.vercel.app,http://localhost:3000"
	}
	allowedOrigins := make(map[string]struct{})
	for _, origin := range strings.Split(allowedOriginsEnv, ",") {
		origin = strings.TrimSpace(origin)
		if origin != "" {
			allowedOrigins[origin] = struct{}{}
		}
	}
	allowCredentials := strings.EqualFold(os.Getenv("CORS_ALLOW_CREDENTIALS"), "true")

	router.Use(func(c *gin.Context) {
		origin := c.Request.Header.Get("Origin")
		if origin != "" {
			if _, ok := allowedOrigins[origin]; ok {
				c.Writer.Header().Set("Access-Control-Allow-Origin", origin)
				c.Writer.Header().Set("Vary", "Origin")
				if allowCredentials {
					c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
				}
			}
		}

		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE, PATCH")

		// Echo requested headers for preflight (more reliable than hardcoding)
		if acrh := c.Request.Header.Get("Access-Control-Request-Headers"); acrh != "" {
			c.Writer.Header().Set("Access-Control-Allow-Headers", acrh)
		} else {
			c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
		}

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	})

	router.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"status":  "ok",
			"message": "Server is running",
		})
	})

	auth := router.Group("/api/auth")
	{
		auth.POST("/register", authHandler.Register)
		auth.POST("/request-code", authHandler.RequestCode)
		auth.POST("/login", authHandler.Login)
		auth.POST("/refresh", authHandler.RefreshToken)
	}

	api := router.Group("/api")
	api.Use(middleware.AuthRequired())
	{
		api.GET("/profile", authHandler.GetProfile)
		api.GET("/organisation", orgHandler.GetOrganisation)
		
		// Availability configuration endpoints
		api.POST("/availability-config", userAvailConfigHandler.CreateConfig)
		api.GET("/availability-config", userAvailConfigHandler.GetConfig)
		api.PUT("/availability-config", userAvailConfigHandler.UpdateConfig)
		api.DELETE("/availability-config", userAvailConfigHandler.DeleteConfig)
		api.GET("/availability-config/check", userAvailConfigHandler.HasConfig)
		
		// Match endpoints for all authenticated users
		api.GET("/matches/current", matchHandler.GetCurrentMatch)
		api.GET("/matches/history", matchHandler.GetMatchHistory)
		api.PATCH("/matches/:id/accept", matchHandler.AcceptMatch)
		api.PATCH("/matches/:id/reject", matchHandler.RejectMatch)
		api.GET("/matches/:id/availabilities", matchHandler.GetMatchAvailabilities)
		
		// Feedback endpoints
		api.POST("/matches/:id/feedback", feedbackHandler.SubmitFeedback)
		api.GET("/matches/:id/feedbacks", feedbackHandler.GetMatchFeedbacks)
		api.GET("/matches/pending-feedback", feedbackHandler.GetPendingFeedback)

		admin := api.Group("/admin")
		admin.Use(middleware.AdminRequired())
		{
			admin.GET("/dashboard", func(c *gin.Context) {
				c.JSON(http.StatusOK, gin.H{
					"message": "Welcome to admin dashboard",
				})
			})
		admin.POST("/import-csv", userHandler.ImportCSV)
		admin.POST("/confirm-user", userHandler.ConfirmUser)
		admin.GET("/users", userHandler.GetOrganisationUsers)
		admin.POST("/users", userHandler.CreateUser)
		admin.DELETE("/users/:id", userHandler.DeleteUser)
		admin.PATCH("/users/:userId/tags", userHandler.UpdateTags)
		admin.GET("/organisation", orgHandler.GetOrganisation)
		admin.PUT("/organisation", orgHandler.UpsertOrganisation)
		
		// Match endpoints for admins
		admin.POST("/matches/generate", matchHandler.GenerateMatches)
		admin.POST("/matches/trigger-scheduler", matchHandler.TriggerScheduler)
		admin.GET("/matches", matchHandler.GetOrganisationMatches)
		admin.GET("/matches/:id/feedbacks", feedbackHandler.AdminGetMatchFeedbacks)
		}
	}

	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	addr := ":" + port
	log.Println("Server starting on " + addr)
	router.Run(addr)
}