# openapi.yaml
openapi: 3.0.3
info:
  title: Virtual Cuppa API
  description: |
    Backend API for Virtual Cuppa application with email-based authentication and automatic matching system.

    ## Features
    - Email-based authentication with 6-digit codes
    - User tagging system for better matching
    - Automatic match generation every Monday at 9 AM
    - Manual match triggering for admins
    - Match acceptance/rejection workflow
    - SendGrid email notifications

    ## Matching System
    The system automatically generates matches between users based on:
    - **Tag compatibility**: Jaccard similarity (common tags / total unique tags × 100)
    - **Availability configuration**: Both users must have availability configuration with at least one common time slot
    - **User status**: No pending matches
    - **Cooldown period**: 30-day minimum between same users
    - **One match per user**: Each user can have only one pending match at a time

    ### Availability Configuration (Required for Matching)
    - Each user must create their **system-wide availability configuration**
    - Configuration defines available days (Monday-Sunday) and periods (morning/afternoon)
    - **Matching requirement**: Both users must have:
      1. Created availability configuration
      2. At least one common time slot (e.g., both available Tuesday afternoon)
    - Users without configuration **will not be matched**

    ### Match Generation Algorithm
...
    4. Sort pairs by score (highest first)
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Register a new Admin account. A confirmation code will be sent to the user's email. After registration, user must login to receive access tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registration successful. Confirmation code sent to email
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/request-code:
    post:
      tags:
        - Authentication
      summary: Request a confirmation code
      description: Request a 6-digit confirmation code to be sent via email (valid for 5 minutes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        "200":
          description: Confirmation code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Confirmation code sent to email
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login with confirmation code
      description: Login using email and 6-digit confirmation code received via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - confirmCode
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                confirmCode:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "123456"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: refresh.token.here
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/profile:
    get:
      tags:
        - User
      summary: Get current user profile
      description: Get the profile of the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/organisation:
    get:
      tags:
        - User
      summary: Get organisation
      description: Get organisation details for the currently logged-in user (both User and Admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Organisation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not assigned to any organisation or organisation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/availability-config:
    post:
      tags:
        - Availability
      summary: Create availability configuration
      description: |
        Create system-wide availability configuration for the authenticated user.
        This configuration defines when the user is available for matches.

        **Required for matching**: Users without availability configuration will not be matched.

        **Matching logic**: A match is created only when both users have availability configuration AND have at least one common time slot.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilityConfigInput"
      responses:
        "201":
          description: Availability configuration created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAvailabilityConfig"
        "400":
          description: Bad request - at least one slot must be selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Configuration already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags:
        - Availability
      summary: Get availability configuration
      description: Get the availability configuration for the authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Availability configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAvailabilityConfig"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Availability
      summary: Update availability configuration
      description: |
        Update availability configuration for the authenticated user.
        Only provided fields will be updated (partial update).
        At least one time slot must remain selected after the update.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityConfigInput"
      responses:
        "200":
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAvailabilityConfig"
        "400":
          description: Bad request - at least one slot must be selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Availability
      summary: Delete availability configuration
      description: Delete the availability configuration for the authenticated user. User will not be matched until a new configuration is created.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Configuration deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Availability configuration deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/availability-config/check:
    get:
      tags:
        - Availability
      summary: Check if user has availability configuration
      description: Check whether the authenticated user has created an availability configuration
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Has configuration status
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasConfig:
                    type: boolean
                    example: true
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/dashboard:
    get:
      tags:
        - Admin
      summary: Admin dashboard
      description: Admin dashboard endpoint (requires admin role)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Welcome to admin dashboard
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/import-csv:
    post:
      tags:
        - Admin
      summary: Import users from CSV
      description: Bulk import users from CSV file (admin only, requires organisation)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with format firstName,lastName,email
      responses:
        "200":
          description: Users imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Users imported successfully
                  count:
                    type: integer
                    example: 10
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/confirm-user:
    post:
      tags:
        - Admin
      summary: Confirm imported user
      description: Confirm a user that was imported via CSV (admin only, same organisation)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: integer
                  example: 5
      responses:
        "200":
          description: User confirmed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User confirmed successfully
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required or wrong organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: Get users by organisation
      description: Get all users belonging to the current user's organisation (admin only)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  count:
                    type: integer
                    example: 10
        "400":
          description: User not assigned to any organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Admin
      summary: Create a new user
      description: Create a new user in the admin's organisation (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - email
              properties:
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request or admin not assigned to organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/users/{id}:
    delete:
      tags:
        - Admin
      summary: Delete a user
      description: Delete a user from the admin's organisation (admin only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID to delete
          example: 5
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted successfully
        "400":
          description: Bad request or user not in same organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/users/{userId}/tags:
    patch:
      tags:
        - Admin
      summary: Update user tags
      description: |
        Update tags for a user in the admin's organisation.

        **Behavior**:
        - Removes all existing tags from the user
        - Creates new tags if they don't exist
        - Assigns all specified tags to the user

        **Use case**: Tags are used for matching algorithm to calculate compatibility scores.
        Users with similar tags will have higher match scores.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: User ID to update tags for
          example: 5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tags
              properties:
                tags:
                  type: array
                  items:
                    type: string
                  description: Array of tag names to assign to the user
                  example: ["developer", "senior", "backend"]
      responses:
        "200":
          description: Tags updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tags updated successfully
        "400":
          description: Bad request or user not in same organisation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/organisation:
    get:
      tags:
        - Admin
      summary: Get organisation
      description: Get organisation details for the currently logged-in user (admin only, alias for backward compatibility)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Organisation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not assigned to any organisation or organisation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Admin
      summary: Upsert organisation
      description: Create or update organisation (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                id:
                  type: integer
                  description: Organisation ID (required for update, omit for create)
                  example: 1
                name:
                  type: string
                  description: Organisation name
                  example: My Company
                companyUrl:
                  type: string
                  description: Company website URL
                  example: https://mycompany.com
      responses:
        "200":
          description: Organisation created or updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organisation"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/matches/current:
    get:
      tags:
        - User
        - Matches
      summary: Get current match
      description: |
        Retrieve the current match for the authenticated user.

        **Current match logic**:
        - Returns matches with status "pending" (waiting for acceptance)
        - Returns matches with status "accepted" IF they haven't expired yet
        - Accepted matches expire 5 days after both users accept
        - After expiration, accepted matches are no longer "current" and move to history
        - Matches where both users have already submitted feedback are no longer considered "current"

        **Returns**:
        - Match details including both users' information
        - **Both users' availability configurations** (availabilityConfig) showing common time slots
        - Match score and scheduled date/time
        - Status ("pending" or "accepted")
        - User tags

        **Note**: Users can only have one current match at a time. After both users submit feedback,
        the system automatically tries to generate a new match for each user.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current pending match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "204":
          description: No pending match found (No Content)
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/matches/history:
    get:
      tags:
        - User
        - Matches
      summary: Get match history
      description: |
        Retrieve complete match history for the authenticated user.

        **Returns**:
        - All matches (pending, accepted, rejected, expired)
        - Chronological order (newest first)
        - Full details of each match including other user's information
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Match history
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      $ref: "#/components/schemas/Match"
                  count:
                    type: integer
                    example: 5
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message description